requirements.txt
feedparser
schedule
python-telegram-bot
openai 